fix getContentRoutes child path and a typo in the prop injector
the prop injector was incorrectly using the contentProp for the fallbackProp!

The changes for the fullpath in the getContentRoutes function wasn't taking into account the fact that we have a first child with a path of ''. This meant the parent would fall back if we had both a name and a fullpath. That in turn caused it to return only 1 in the route stack instead of two. And that would make the parent's content show up on the child.