Added wrapper classes for many levels of the content and tlazy-loaded the children and fixed some bugs.
Made a lot of structure changes. I wanted to make them anyway, but the motivation was to make it easier to use this for another project that uses it.

Added new route and prop clases
added accessores to these objects and the content class to lazy-parse the routes and their props as-needed.
moved the injection code from the index.js to the respective level in the content -> route -> prop hierarchy.
While doing this I fxed a bug with the getContentRoute where it would return null (formally []) when it didn't match against the first child.